XXX: Disable allocation deadline
Plane allocation takes a really long time (tens to hundreds of ms) on a
Pi 4 because we have a decent number of planes and the algorithm is
combinatorial.  So we basically always hit the timeout and just do
composition.  This makes things more usable but also isn't very
interesting.  So for now just disable the timeout so we do actually
allocate hardware planes, and at some point I'll improve the allocator
to go faster.