Force primary-composition allocation when compositing
When doing composition, it will not be valid to allocate the composition
layer to any plane other than the primary.  We can't trivially work out
when composition is required or not, but if there are more (visible,
non-composition) layers than planes then we are guaranteed to need
composition.  In this case we know that any branches where the primary
plane is allocated to a layer other than the composition layer are
dead-ends, so we can disallow such an allocation to save time.